requirements.txt
click==8.1.7
cmake==3.27.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
mysql-connector-python==8.2.0
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
protobuf==4.21.12